requirements.txt
pytest>=6.1.2
discord.py>=1.5.1
python-dotenv>=0.15.0